Added an option, based in spirit on the patches from Chris Farber, which make the status item optional.
If you turn off "Show in Menu Bar", Scrollvetica will suppress the status
item when it is launched as a login item. (This should cover nearly all of
the cases where people want to hide the status item, while still making it
behave reasonably for quick interactive use and trying out.)

If the status item is hidden, double-clicking the Scrollvetica icon in the
finder will make it visible again.